requirements.txt
streamlit
sentence-transformers
faiss-cpu
scikit-learn
pandas
numpy